Create shared `conftest.py` for PFS/MS/libs
