+++ tests/libs/conftest.py
@@ -1,10 +1,6 @@
-from gevent import monkey
-
 from raiden_contracts.tests.fixtures import *  # noqa
 from raiden_libs.test.fixtures.address import *  # noqa
 from raiden_libs.test.fixtures.client import *  # noqa
 from raiden_libs.test.fixtures.web3 import *  # noqa
 
 from .fixture_overwrites import *  # noqa
-
-monkey.patch_all()
